requirements.txt
aiogram==2.25.2
APScheduler==3.10.4
beautifulsoup4==4.12.3
googletrans==4.0.0rc1
instaloader==4.13.1
Pillow==10.4.0
pinterest_api==0.0.8
pypdl==1.3.2
Requests==2.32.3
yt_dlp==2024.8.6
